modified category/subcategory/material endpoints
