+++ Client/src/Components/Search/SearchDatasetsForm.tsx
@@ -3,7 +3,10 @@ import { Field, FieldArray, Form, Formik } from 'formik'
 import { ICategory, ISearchDatasetsFormModel, defaultSearchDatasetsModel, searchDatasetsValidationSchema } from './ISearchDatasetsFormModel'
 import { MuiSelectFormik, MuiTextFieldFormik } from '../Forms/FormikFields'
 import React, { useEffect, useState } from 'react'
-import { listCategories, listMaterials, listSubcategories } from '../../Remote/Endpoints/DatasetEndpoint'
+import { listMaterials } from '../../Remote/Endpoints/materialEndpoint'
+import { listCategories } from '../../Remote/Endpoints/categoryEndpoint'
+import { listSubcategories } from '../../Remote/Endpoints/subcategoryEndpoint'
+
 
 import { MaterialSelectChipArray } from '../DatasetUpload/MetaSection/MaterialSelectChipArray'
