+++ Client/src/Remote/Endpoints/DatasetEndpoint.tsx
@@ -15,29 +15,29 @@ export const getDatasets = async (query: ISearchDatasetsFormModel): Promise<IDat
   return localDatasets
 }
 
-interface ICategoryModel {
-  id: number,
-  name: string
-}
-interface ISubcategoryModel {
-  id: number,
-  name: string
-}
-interface IMaterialModel {
-  details: string,
-  composition: string,
-  id: number
-}
+// interface ICategoryModel {
+//   id: number,
+//   name: string
+// }
+// interface ISubcategoryModel {
+//   id: number,
+//   name: string
+// }
+// interface IMaterialModel {
+//   details: string,
+//   composition: string,
+//   id: number
+// }
 
-export const listCategories = async (): Promise<ICategoryModel[]> => {
-  const categories = await get('/category')
-  return categories
-}
-export const listSubcategories = async (): Promise<ISubcategoryModel[]> => {
-  const subcategories = await get('/subcategory')
-  return subcategories
-}
-export const listMaterials = async (): Promise<IMaterialModel[]> => {
-  const materials = await get('/material')
-  return materials
-}
+// export const listCategories = async (): Promise<ICategoryModel[]> => {
+//   const categories = await get('/category')
+//   return categories
+// }
+// export const listSubcategories = async (): Promise<ISubcategoryModel[]> => {
+//   const subcategories = await get('/subcategory')
+//   return subcategories
+// }
+// export const listMaterials = async (): Promise<IMaterialModel[]> => {
+//   const materials = await get('/material')
+//   return materials
+// }
