+++ Client/src/Remote/Endpoints/categoryEndpoint.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import { get } from "../RemoteHelper"
+
+
+interface ICategoryModel {
+    id: number,
+    name: string
+}
+export const listCategories = async (): Promise<ICategoryModel[]> => {
+    const categories = await get('/category')
+    return categories
+}
