+++ Client/src/Remote/Endpoints/materialEndpoint.tsx
@@ -1,0 +1,12 @@
+import { get } from "../RemoteHelper"
+interface IMaterialModel {
+    details: string,
+    composition: string,
+    id: number
+}
+
+
+export const listMaterials = async (): Promise<IMaterialModel[]> => {
+    const materials = await get('/material')
+    return materials
+}
