+++ Client/src/Remote/Endpoints/subcategoryEndpoint.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import { get } from "../RemoteHelper"
+
+interface ISubcategoryModel {
+    id: number,
+    name: string
+}
+
+export const listSubcategories = async (): Promise<ISubcategoryModel[]> => {
+    const subcategories = await get('/subcategory')
+    return subcategories
+}
